Dltsdev

+++ src/init-scripts/bootstrap.sh
@@ -3,23 +3,32 @@ set -ex
 
 SCRIPT_DIR=/pod/scripts
 
-# Dir for saving running status
-export PROC_DIR=/pod/running
-rm -rf ${PROC_DIR}
-mkdir -p ${PROC_DIR}
+echo bootstrap starts at `date` &>> ${LOG_DIR}/bootstrap.log
 
-# Dir for logs
-export LOG_DIR=/pod/logs
-rm -rf ${LOG_DIR}
-mkdir -p ${LOG_DIR}
+if [ "$DLWS_ROLE_NAME" != "inferenceworker" ];
+then
+	# Dir for saving running status
+	export PROC_DIR=/pod/running
+	rm -rf ${PROC_DIR}
+	mkdir -p ${PROC_DIR}
 
-# Save the pid.
-PID_FILE=${PROC_DIR}/pid
-echo $$ > $PID_FILE
+	# Dir for logs
+	export LOG_DIR=/pod/logs
+	rm -rf ${LOG_DIR}
+	mkdir -p ${LOG_DIR}
+
+	# Save the pid.
+	PID_FILE=${PROC_DIR}/pid
+	echo $$ > $PID_FILE
+fi
 
 # Setup container
 bash ${SCRIPT_DIR}/init_user.sh &>> ${LOG_DIR}/bootstrap.log
-touch ${PROC_DIR}/CONTAINER_READY
+
+if [ "$DLWS_ROLE_NAME" != "inferenceworker" ];
+then
+	touch ${PROC_DIR}/CONTAINER_READY
+fi
 
 # Setup roles
 bash ${SCRIPT_DIR}/setup_sshd.sh &>> ${LOG_DIR}/bootstrap.log
@@ -29,11 +38,16 @@ then
     bash ${SCRIPT_DIR}/setup_ssh_config.sh &>> ${LOG_DIR}/bootstrap.log
 fi
 
-touch ${PROC_DIR}/ROLE_READY
+if [ "$DLWS_ROLE_NAME" != "inferenceworker" ];
+then
+	touch ${PROC_DIR}/ROLE_READY
 
-# Setup job
-# TODO
-touch ${PROC_DIR}/JOB_READY
+	# Setup job
+	# TODO
+	touch ${PROC_DIR}/JOB_READY
+fi
+
+echo bootstrap ends at `date` &>> ${LOG_DIR}/bootstrap.log
 
 set +e
 # Execute user's command for the job
